test: 테스트 코드 기능 별 병합
